backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.28
psycopg2-binary==2.9.9
pydantic==2.6.4
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.9
confluent-kafka==2.3.0
fastavro==1.9.4
requests==2.31.0
Pillow>=10.4.0
aiofiles==23.2.1
python-dotenv==1.0.1
alembic==1.13.1
greenlet==3.0.3
email-validator==2.1.1
